fix after jdt.ls refactoring
